Several corrections to lib/Sitemap.
Fixed grammar of several titles, and changed wording of a few headings
to avoid redundancy, which sounds bad in English - for instance, saying
"adding additional". Fixed capitalisation of titles following normative
editorial rules: all words should be capitalised, except prepositions
(to, with, on), indefinite articles (a, an) and names which are normally
spelled with lowercase letters.